refactor: consolidate validation functions into validators module
