+++ app/api/company.py
@@ -44,7 +44,7 @@ from app.src.regex import NAME_PATTERN
 from app.src.enums import CompanyStatus, CompanyType, OrderIn
 from app.src.urls import URL_COMPANY
 from app.src.openobserve import log_event
-from app.src.validators import validate_id, verify_permission, verify_token
+from app.src.validators import validate_id, verify_permission, verify_token, validate_srid_4326, validate_wkt_string
 from app.src.functions import (
     apply_created_on_filters,
     apply_updated_on_filters,
@@ -56,8 +56,7 @@ from app.src.functions import (
     get_operator_roles,
     get_request_info,
     update_if_changed,
-    validate_srid_4326,
-    validate_wkt_string,
+    
     resolve_model_defaults,
     apply_status_filters,
     apply_type_filters,
